lab4-step2: MVC warehouse version

+++ lab_4/reactive/src/main/kotlin/it/polito/ecommerce/warehouse/domain/Product.kt
@@ -9,5 +9,5 @@ data class Product (
     val name: String,
     val category: String,
     val price: BigDecimal,
-    var quantity: Int
+    var quantity: Long
 )
